[WIP]bad_connector_woocommerce:create binded product_template and import all variant products
